예외처리 수정

+++ src/main/java/com/hanghae/my_blog/service/CommentService.java
@@ -6,6 +6,7 @@ import com.hanghae.my_blog.dto.CompleteResponseDto;
 import com.hanghae.my_blog.entity.Comment;
 import com.hanghae.my_blog.entity.Post;
 import com.hanghae.my_blog.entity.User;
+import com.hanghae.my_blog.entity.UserRoleEnum;
 import com.hanghae.my_blog.jwt.JwtUtil;
 import com.hanghae.my_blog.repository.CommentRepository;
 import com.hanghae.my_blog.repository.PostRepository;
@@ -26,13 +27,14 @@ public class CommentService {
     private final CommentRepository commentRepository;
     private final UserRepository userRepository;
 
+    @Transactional
     //댓글 저장하기
     public CommentResponseDto saveComment(Long postid, CommentRequestDto commentRequestDto, HttpServletRequest request) {
         //로그인 여부 확인
         User user = tokenChecking(request);
         //게시글 저장 여부 확인
         Post post = postRepository.findById(postid).orElseThrow(
-                () -> new NullPointerException("게시글이 존재하지 않습니다")
+                () -> new IllegalArgumentException("게시글이 존재하지 않습니다")
         );
         //댓글 저장
         Comment comment = new Comment(commentRequestDto,  post, user);
@@ -44,32 +46,41 @@ public class CommentService {
         //로그인 여부 확인
         User user = tokenChecking(request);
         //게시글 존재 여부 확인
-        Post post = postRepository.findById(postid).orElseThrow(
-                () -> new NullPointerException(("게시글이 존재하지 않습니다")
+        postRepository.findById(postid).orElseThrow(
+                () -> new IllegalArgumentException(("게시글이 존재하지 않습니다")
         ));
         //댓글 존재 여부 확인
         Comment comment = commentRepository.findById(commentid).orElseThrow(
-                () -> new NullPointerException(("댓글이 존재하지 않습니다")
+                () -> new IllegalArgumentException(("댓글이 존재하지 않습니다")
         ));
-        //댓글 업데이트
+        //ADMIN 권한 또는 username 확인 후 댓글 업데이트
+        if(comment.getUser().getUsername().equals(user.getUsername()) || user.getRole() == UserRoleEnum.ADMIN){
         comment.update(commentRequestDto);
+        }else{
+            throw new IllegalArgumentException("올바른 사용자가 아닙니다");
+        }
         //수정된 댓글 반환
         return new CommentResponseDto(comment.getContent());
     }
+    @Transactional
     //댓글 삭제하기
     public CompleteResponseDto deleteComment(Long postid, Long commentid, HttpServletRequest request) {
         //로그인 여부 확인
         User user = tokenChecking(request);
         //게시글 저장 여부 확인
-        Post post = postRepository.findById(postid).orElseThrow(
-                () -> new NullPointerException(("게시글이 존재하지 않습니다")
+        postRepository.findById(postid).orElseThrow(
+                () -> new IllegalArgumentException(("게시글이 존재하지 않습니다")
         ));
         //댓글 저장 여부 확인
         Comment comment = commentRepository.findById(commentid).orElseThrow(
-                () -> new NullPointerException(("댓글이 존재하지 않습니다")
+                () -> new IllegalArgumentException(("댓글이 존재하지 않습니다")
         ));
-        //댓글 삭제
-        commentRepository.delete(comment);
+        //ADMIN 권한 또는 username 확인 후 댓글 삭제
+        if(comment.getUser().getUsername().equals(user.getUsername()) || user.getRole() == UserRoleEnum.ADMIN){
+            commentRepository.delete(comment);
+        }else{
+            throw new IllegalArgumentException("올바른 사용자가 아닙니다");
+        }
         //삭제 완료 반환
         return new CompleteResponseDto("삭제 완료");
     }
@@ -84,7 +95,7 @@ public class CommentService {
                 throw new IllegalArgumentException("Token Error");
             }
             User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
-                    () -> new NullPointerException(("사용자가 존재하지 않습니다")
+                    () -> new IllegalArgumentException(("사용자가 존재하지 않습니다")
             ));
             return user;
         }throw new IllegalArgumentException("로그인이 필요합니다");
